requirements.txt
flask==3.0.2
flask-cors==5.0.0
werkzeug==3.1.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.5.2
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
keras==3.8.0
nltk==3.8.1
joblib==1.4.2
